Move db fixtures to session scope
